+++ network-grpc/src/service/response_future.rs
@@ -1,0 +1,79 @@
+use crate::convert::{error_into_grpc, IntoProtobuf};
+
+use network_core::error as core_error;
+
+use futures::prelude::*;
+use tower_grpc::{self, Code, Status};
+
+use std::{marker::PhantomData, mem};
+
+#[must_use = "futures do nothing unless polled"]
+pub enum ResponseFuture<T, F> {
+    Pending(F),
+    Failed(Status),
+    Finished(PhantomData<T>),
+}
+
+impl<T, F> ResponseFuture<T, F>
+where
+    F: Future,
+    F::Item: IntoProtobuf<T>,
+{
+    pub fn new(future: F) -> Self {
+        ResponseFuture::Pending(future)
+    }
+}
+
+impl<T, F> ResponseFuture<T, F> {
+    pub fn error(status: Status) -> Self {
+        ResponseFuture::Failed(status)
+    }
+
+    pub fn unimplemented() -> Self {
+        ResponseFuture::Failed(Status::new(Code::Unimplemented, "not implemented"))
+    }
+}
+
+impl<T, F> Future for ResponseFuture<T, F>
+where
+    F: Future<Error = core_error::Error>,
+    F::Item: IntoProtobuf<T>,
+{
+    type Item = tower_grpc::Response<T>;
+    type Error = tower_grpc::Status;
+
+    fn poll(&mut self) -> Poll<Self::Item, tower_grpc::Status> {
+        if let ResponseFuture::Pending(f) = self {
+            let res = poll_and_convert_response(f);
+            if let Ok(Async::NotReady) = res {
+                return Ok(Async::NotReady);
+            }
+            *self = ResponseFuture::Finished(PhantomData);
+            res
+        } else {
+            match mem::replace(self, ResponseFuture::Finished(PhantomData)) {
+                ResponseFuture::Pending(_) => unreachable!(),
+                ResponseFuture::Failed(status) => Err(status),
+                ResponseFuture::Finished(_) => panic!("polled a finished response"),
+            }
+        }
+    }
+}
+
+fn poll_and_convert_response<T, F>(
+    future: &mut F,
+) -> Poll<tower_grpc::Response<T>, tower_grpc::Status>
+where
+    F: Future<Error = core_error::Error>,
+    F::Item: IntoProtobuf<T>,
+{
+    match future.poll() {
+        Ok(Async::NotReady) => Ok(Async::NotReady),
+        Ok(Async::Ready(res)) => {
+            let item = res.into_message()?;
+            let response = tower_grpc::Response::new(item);
+            Ok(Async::Ready(response))
+        }
+        Err(e) => Err(error_into_grpc(e)),
+    }
+}
